luci-theme-*: Autogenerate hashes for the static files of LuCI themes
For all static files of the LuCI themes the hash is calculated. A parameter
with a hash of the file is added to the file loading location (in header.htm
and footer.htm) to avoid cache problems when theme is updated. Additionally
added hashes for cbi.js and xhr.js files loading.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>