procd: Add initial cgroup support
Added the ability to change the cgroup at service startup. The required
cgroup and controllers (if needed) must be specified in the init script with
the 'cg_group' parameter. Parameter 'cg_group' has '<controllers>:<path>'
format. For all available controllers '<controllers>' must be set to '*'
or completely omit the '<controller>:' part (with the colon). Parameter
'cg_group' is array and can be set multiple times with procd_append_param.

Also you can set 'cg_sticky' parameter in order to
cgred daemon does not change pidlist and children tasks of the service.

Example 1:

    start_service() {
        procd_open_instance
        procd_set_param command "$BIN"
        procd_set_param cg_group "*:openwrt/high/"
        procd_set_param cg_sticky "1"
        procd_close_instance
    }

Different cgroups for different controllers can be set as follows:

    ...
    procd_set_param cg_group "cpu:openwrt/cpu-high/"
    procd_append_param cg_group "memory:openwrt/mem-low/"
    ...

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>