fstools: Fix block devices hotplug mount/unmount
In commit 73a6f44, the procd-hotplug functionality has been completely
removed. Instead of procd-hotplug, udev is used. The device name in the
DEVICE environment variable from udev comes already with the correct
prefix '/dev/'. Not need to add it to the beginning of the device name.

As a result of this difference in behavior between procd-hotplug and
udev, there were problems with mounting and unmounting media when
connected/disconnected to device.

Fixes: 73a6f44
Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>
(cherry picked from commit c90a7f67f25bffd7916352c98de2d7e4044660be)