procd: inittab: Handle multiple consoles in /proc/cmdline
If askconsole (or askconsolelate) specified without id, then process
will be executed for all tty's defined in the kernel command line
using "console=" parameters.

If askconsole (or askconsolelate) has specified id, then process will
be executed only for tty specified in the id, if this tty is also
defined in the command line using "console=" parameter.

The sequence of starting inittab actions:
  1. respawn
  2. askconsole
  3. askfirst
  4. sysinit
  5. respawnlate
  6. askconsolelate
  7. shutdown

If a process has already been started for tty through the askconsole
action, then the askfirst entries for the same tty will be ignored.

If a process has already been started for tty through the askfirst
action, then no process will be started for this tty through the
askconsolelate action.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>