ddns-scripts: Fix slow scripts execution
The dynamic_dns_functions.sh script searches for libcurl.so and
libustream-ssl.so libraries in the /lib and /usr/lib folders and
all their subfolders. If these folders have a lot of subfolders
attached, it takes a very long time. To increase the search speed,
this commit limits the maximum search depth to 1.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>