collectd: Fix random crashes at startup
In commit 55e7589 the used version of collectd was downgraded to 5.8.1
due to an random crashes of collectd 5.10.0 starting. However, it was
found that similar crashed are also present in version 5.8.1.

Detailed research has resulted in crashes that occurred after downgrading
the used version of rrdtool to 1.0.50 (commits 796bf19 and 31fed43).
One of the main differences is that rrdtool 1.0.50 is not threadsafe.

As it turned out, this is what causes the crashes. The collectd code used
two different locks (mutexes) for the locking rrdtool library for calls
rrd_create() and rrd_update(). The correct solution in this case is to
using one lock for all rrdtool library calls.

Refs: 55e7589 ("collectd: Prefer version 5.8.1 (downgrade)")
Refs: 796bf19 ("distro: Set preferred version for rrdtool to 1.0.50")
Refs: 31fed43 ("rrdtool: Add recipe for version 1.0.50")
Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>