base-files: Check and try to fix /etc/passwd and /etc/shadow at boot
If we have any problems with /etc/passwd, the ubus service will not be
able to determine the credentials (the glibc getpwuid() and getpwgrid()
functions are used) of the client and will refuse to connect any client.
The first client to connect to ubus is procd. If procd fails to connect
to ubus, the device will turn into a brick. We try to prevent this by
recovering files from ROM in case they are damaged.

Such failures with /etc/passwd have been reported on some devices with
overlayfs in case of power failure immediately after password change
by passwd or chpasswd utilities.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>