luci-theme-tano: Fix HOME location
By default the NPM looks in the user's home directory for a .npmrc
file to read the configuration. If the user who started the build
has the .npmrc file available, unexpected problems may occur.
To avoid such problems, we redefine the current home directory
to the working directory of the package build.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>