base-files: Dynamic values substitution on banner displaying
Any variables from '/etc/openwrt_release' and/or '/etc/device_info' files
can be used in the '/etc/issue' file in the format '${VAR}'. When the banner
is displayed (at login), all occurrences of '${VAR}' will be automatically
replaced with real values. The replacement of variables is performed
in the '/etc/profile' script using sed.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>