tanowrt-version.bbclass: Drop variables with timestamps and revisions
The following variables and their substitution by the OPENWRT_VERSION_SED
script have been removed:
  - OPENWRT_VERSION_SRCREV
  - OPENWRT_VERSION_SRCREV_SHORT
  - OPENWRT_VERSION_CODE
  - OPENWRT_VERSION_CODE_LU

Using this variables may cause problems with siggens when using esdk.

The inheritance of the tano-scmrevision class was removed along with
the variables (it's no longer needed here because there are no more
variables with revisions).

In addition, the default values for some variables have been corrected.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>