treewide: Remove useless tanowrt class
This class contains nothing except FILES_SOLIBSDEV definition
and tanowrt-lua class inheritance.

Completely remove this class, replacing its inheritance in recipes
with the inheritance of the tanowrt-lua class and add FILES_SOLIBSDEV
where necessary.

Also, in some recipes where the class tanowrt-lua is not needed,
it was necessary to add "${libdir}" to "FILES" and fix the paths
in CMakeLists.txt (rpcd).

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>