base-files: sysfixtime: Set localtime by newest file from /etc
When booting the system, we calculate the maximum time available from
files in the /etc folder. If the time in RTC is less than the calculated
time, we set the local time to the calculated value.

This allows you to avoid setting the local time to various fantastic
values, such as 1970, etc., and also helps to bring the local time as
close to real values as possible.

Also slightly improved error handling and logging in sysfixtime script.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>