ovmf: Set PACKAGE_ARCH to MACHINE_ARCH
This was required because the package generates the blobs to the
deploy directory when it is built. For example, ovmf.qcom2 and
other files is generated when building for qemux86-64 machine.
If we keep the PACKAGE_ARCH according to the SoC architecture,
then when building package for another machine, the required blob
may not appear in the deploy directory, because no rebuild of the
package will be required.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>