layer: Provide files with licenses for required recipes
In the hardknott release GPL-2.0 license file in common files
are renamed to GPL-2.0-only and to GPL-2.0-or-later. If we refer
in LIC_FILES_CHKSUM to the GPL-2.0 file in the common licenses
folder, this does not work in the hardknott release.

To fix this issue and ensure compatibility when upgrading to
a hardknott release, provide the required license file for each
recipe, and refer to it in LIC_FILES_CHKSUM.

This commit is a small preparation for the hardknott release.

Provided files with licenses for recipes:
  - wwan (GPL-2.0)
  - usbreset (GPL-2.0)
  - procd-inittab (GPL-2.0)
  - resolveip (GPL-2.0)
  - speedtest-netperf (GPL-2.0)
  - urandom-seed (GPL-2.0)
  - qt5-env (GPL-2.0)
  - qos-sciprts (GPL-2.0)
  - proto-vxlan (GPL-2.0)
  - proto-gre (GPL-2.0)
  - proto-bonding (GPL-2.0)
  - mwan3 (GPL-2.0)
  - ddns-scripts (GPL-2.0 and MPL-2.0)
  - cfg80211 (GPL-2.0)

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>