tanowrt-kmods: Fallback kernel version detection to filesystem method
The kernel version using the get_kernelversion_headers() function cannot
be determined if STAGING_KERNEL_BUILDDIR is not populated. This can
happen, for example, if the kernel was not built, but taken from
sstate-cache. For such cases, the rootfs_flatten_modules_hook()
function will determine the kernel version directly from the directory
name on the filesystem in /lib/modules. We could add a dependency on
virtual/kernel:do_shared_workdir task, but in this case we would have
to always do a full kernel rebuild to populate the rootfs.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>