kernel-tano: More intelligent default value for LINUX_LOCALVERSION
Using git to find the current revision of the kernel source
code may return a revision which does not match the version
of the package due to applied patches. As a result, the version
in include/generated/utsrelease.h and the version of the kernel
packages will be different. To avoid this issues, use more
intelligent method for default value for LINUX_LOCALVERSION
with using KERNEL_SRC_SRCREV value if it is not empty.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>
(cherry picked from commit dfce7ab36df054c52d25d7d6aad32bcd6e7a9de0)